[rename] folder structure changed
