--- client/src/components/mypage/Bio.module.css
@@ -1,8 +1,0 @@
-.bioContainer{
-    padding: 24px;
-}
-
-.bioContent{
-    font-size: 0.8rem;
-    color: var(--color-gray-1);
-}
