--- client/src/components/mypage/CreateProfile.module.css
@@ -1,22 +1,0 @@
-.createForm {
-  display: flex;
-  flex-direction: column;
-  gap: 80px;
-  width: 100%;
-  background-color: white;
-}
-
-
-.formItem {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.helpText {
-  font-size: 0.8rem;
-}
-
-.warning {
-  color: var(--color-state-error);
-}
