--- client/src/components/mypage/DropDownTag.module.css
@@ -1,59 +1,0 @@
-.dropDownTagContainer {
-  position: relative;
-  display: flex;
-  gap: 8px;
-  border-radius: 24px;
-  width: fit-content;
-  border: 1px solid var(--color-gray-3);
-  padding: 4px 14px;
-  cursor: pointer;
-}
-
-.tagName {
-  color: var(--color-gray-3);
-}
-
-.levelDropdown {
-  /* position: relative; */
-  /* padding: 4px 8px; */
-  /* border: 1px solid var(--color-gray-3); */
-  cursor: pointer;
-}
-
-.optionsContainer {
-  position: absolute;
-  top: 100%;
-  left: 0;
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  border: 1px solid var(--color-gray-3);
-  border-radius: 16px;
-  background-color: white;
-  z-index: 10;
-}
-
-.optionsContainer .option {
-  /* padding: 8px; */
-  cursor: pointer;
-  text-align: center;
-}
-
-.optionsContainer .option:hover {
-  background-color: var(--color-sub);
-  border-radius: 40px;
-}
-
-.hidden {
-  display: none;
-}
-
-
-.active{
-  background-color: var(--color-main);
-  border: none;
-}
-
-.active span {
-  color: white;
-}
