--- client/src/components/mypage/EditInfo.module.css
@@ -1,54 +1,0 @@
-.editformContainer {
-  display: flex;
-  /* justify-content: center; */
-  align-items: center;
-}
-
-.editform {
-  width: 60%;
-  padding: 24px;
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.edititem {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.labelContainer {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  gap: 8px;
-}
-
-.helpText {
-  text-align: right;
-  font-size: 0.8rem;
-  color: var(--color-state-error);
-}
-
-.edititem label {
-  min-width: fit-content;
-}
-
-.edititem input {
-  width: 70%;
-}
-
-.actions {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 8px;
-}
-
-.save {
-  background-color: var(--color-main);
-  color: white;
-  border: none;
-  cursor: pointer;
-}
