--- client/src/components/mypage/HardInput.module.css
@@ -1,7 +1,0 @@
-.hardInputContainer{
-}
-
-.hardInput{
-    border-radius: 50px;
-    width: 100%;
-}
