--- client/src/components/mypage/HardProfile.module.css
@@ -1,35 +1,0 @@
-.hardContainer {
-  display: flex;
-  align-items: center;
-  margin-top: 16px;
-}
-
-.tagContainer {
-  display: flex;
-  gap: 8px;
-  border-radius: 24px;
-  width: max-content;
-  /* height: ; */
-  padding: 4px 14px;
-  background-color: var(--color-main);
-  border: none;
-  cursor: pointer;
-}
-
-.tagContainer span {
-  color: white;
-}
-
-.level {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background-color: var(--color-sub);
-  width: 100px;
-  box-shadow: 0 0 10px 0 rgba(0, 0, 0, 0.3);
-}
-
-.level span {
-  color: var(--color-main);
-  font-weight: 600;
-}
