--- client/src/components/mypage/NoContent.module.css
@@ -1,19 +1,0 @@
-.noContentContainer{
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    height: 100px;
-    margin-top: 2rem;
-}
-
-.facelogo{
-    width: 80px;
-    height: auto;
-    margin-bottom: 1rem;
-    opacity: 0.8;
-}
-
-.helpText{
-    color: var(--color-gray-3);
-}
