--- client/src/components/mypage/ProfileCats.tsx
@@ -1,9 +1,0 @@
-enum ProfileCats {
-  BIO = "저는 이런 사람이에요.",
-  TECH = "기술 스택",
-  HARD = "하드 스킬",
-  SOFT = "소프트 스킬",
-  PROJ = "참여한 프로젝트",
-}
-
-export default ProfileCats;
