--- client/src/components/mypage/ProjCard.module.css
@@ -1,23 +1,0 @@
-.cardContainer {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 8px;
-  /* 임시 */
-  width: 180px;
-  margin-top: 16px;
-}
-
-.projImg {
-  border-radius: 20px;
-  background-color: var(--color-main);
-  width: 100%;
-  height: auto;
-  object-fit: cover;
-  box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.2);
-}
-
-.projTitle {
-  font-size: 1rem;
-  color: var(--color-gray-1);
-}
