--- client/src/components/mypage/SoftInput.module.css
@@ -1,7 +1,0 @@
-.softInputContainer{
-}
-
-.softInput{
-    border-radius: 50px;
-    width: 100%;
-}
