--- client/src/components/mypage/SoftTag.module.css
@@ -1,14 +1,0 @@
-.tagContainer {
-  display: flex;
-  gap: 8px;
-  border-radius: 24px;
-  width: fit-content;
-  border: 1px solid var(--color-gray-3);
-  padding: 4px 14px;
-  background-color: var(--color-main);
-  border: none;
-}
-
-.tagContainer span {
-  color: white;
-}
