--- client/src/components/mypage/Tag.module.css
@@ -1,26 +1,0 @@
-.tagContainer{
-    display: flex;
-    gap: 8px;
-    border-radius: 24px;
-    width: fit-content;
-    border: 1px solid var(--color-gray-3);
-    padding: 4px 14px;
-}
-
-.tagName{
-    color: var(--color-gray-3);
-}
-
-.tagContainer:hover{
-    background-color: var(--color-sub);
-}
-
-.tagContainer.active {
-    background-color: var(--color-main);
-    border: none;
-}
-
-
-.tagContainer.active span {
-    color: white;
-}
