--- client/src/components/mypage/TagRemover.tsx
@@ -1,10 +1,0 @@
-import { FC } from "react";
-
-interface Props {
-  content: string;
-}
-const TagRemover: FC<Props> = ({ content }) => {
-  return <div dangerouslySetInnerHTML={{ __html: content }} />;
-};
-
-export default TagRemover;
