--- client/src/components/mypage/TechTags.module.css
@@ -1,41 +1,0 @@
-.techContainer{
-  display: flex;
-  flex-wrap: wrap;
-  gap: 4px;
-}
-
-.tagsContainer {
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-}
-
-.tagTypeContainer{
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-}
-
-.tagType{
-  font-weight: 500;
-  color: var(--color-gray-1);
-}
-
-.tags{
-  display: flex;
-  flex-wrap: wrap;
-  gap: 4px
-}
-
-.viewMoreContainer {
-  display: flex;
-  gap: 8px;
-  border-radius: 24px;
-  width: max-content;
-  border: 1px solid var(--color-gray-3);
-  padding: 4px 14px;
-}
-
-.viewMore {
-  color: var(--color-gray-3);
-}
