--- client/src/components/mypage/TitleLine.module.css
@@ -1,20 +1,0 @@
-.titleline {
-    display: flex;
-    flex-basis: 100%;
-    align-items: center;
-  }
-  
-  .titleline::after {
-    content: "";
-    flex-grow: 1;
-    background-color: #679c97;
-    height: 1px;
-    font-size: 0px;
-    line-height: 0px;
-    margin-left: 1rem;
-  }
-  
-  .title {
-    font-size: 24px;
-    font-weight: 600;
-  }
