+++ client/src/pages/mypage/Profile.tsx
@@ -2,11 +2,11 @@ import { FC, useEffect, useState } from "react";
 import classes from "./Profile.module.css";
 import NoContent from "../../components/mypage/NoContent";
 import editicon from "../../assets/icons/edit.svg";
-import TagRemover from "../../components/mypage/TagRemover";
+import TagRemover from "../../components/mypage/format/TagRemover";
 import TitleLine from "../../components/mypage/TitleLine";
-import ProfileCats from "../../components/mypage/ProfileCats";
+import ProfileCats from "../../components/mypage/format/ProfileCats";
 import { useNavigate, useParams } from "react-router-dom";
-import ProjCard from "../../components/mypage/ProjCard";
+import ProjCard from "../../components/mypage/view/ProjCard";
 import SoftTag from "../../components/mypage/SoftTag";
 import TechProfile from "../../components/mypage/TechProfile";
 import HardProfile from "../../components/mypage/HardProfile";
